[inductor] Share reduction combine_fn between backends
Currently reductions are hard-coded separately in `ir.py`, `codegen/triton.py`
and `codegen/cpp.py`. This reduces some of that duplication by factoring out the
`combine_fn` definition used in `_unroll_reduction_fn` and using it to generate
some of the reduction code in for the triton and cpp backends.

ghstack-source-id: c3c30d6a5fe8c00a3b9b6bf1c5b5894f907603d9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/99661